jailbroken_ai/requirements.txt
streamlit
ollama
pyyaml
python-dotenv
